requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.4
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
passlib==1.7.4
requests==2.31.0
